% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiral_pkg_versions.R
\name{admiral_pkg_versions}
\alias{admiral_pkg_versions}
\title{Report Package Versions}
\usage{
admiral_pkg_versions()
}
\description{
Function returns a string listing the packages and their versions.
}
\examples{
admiral_pkg_versions()
}
