% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fns_with_keyword.R
\name{get_fns_with_keyword}
\alias{get_fns_with_keyword}
\title{Find Functions with Keyword}
\usage{
get_fns_with_keyword(package, keyword)
}
\arguments{
\item{package}{a single package name}

\item{keyword}{string of the keyword to identify}
}
\value{
a character vector
}
\description{
Returns a character vector of functions that have the passed keyword
in the help file. For example, this function can be used to find all
deprecated or superseded functions in the admiral universe, as admiral
package include \verb{#' @keywords deprecated} or \verb{#' @keywords superseded}
in the function's roxygen2 comments.
}
\examples{
get_fns_with_keyword(package = "admiral", keyword = "deprecated")
}
