% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fns_with_keyword.R
\name{get_keyword_fns}
\alias{get_keyword_fns}
\alias{get_admiral_deprecated}
\alias{get_admiral_superseded}
\alias{get_fns_with_keyword}
\title{Find Functions with Keyword}
\usage{
get_admiral_deprecated()

get_admiral_superseded()

get_fns_with_keyword(package, keyword, lib.loc = NULL)
}
\arguments{
\item{package}{a character string naming an installed package.}

\item{keyword}{string of the keyword to identify}

\item{lib.loc}{a character vector of directory names of \R libraries,
    or \code{NULL}.  The default value of \code{NULL} corresponds to all
    libraries currently known.  The specified library trees are used to
    search for \code{package}.}
}
\value{
a character vector of function names
}
\description{
\code{get_admiral_deprecated()}: Returns tibble of all deprecated functions in
the admiral, admiralonco, admiralophtha, and admiralvaccine packages.

\code{get_admiral_superseded()}: Returns tibble of all superseded functions in
the admiral, admiralonco, admiralophtha, and admiralvaccine packages.

\code{get_fns_with_keyword()}: Returns a character vector of functions that have
the passed keyword in the help file. For example, this function can be used to find all
deprecated or superseded functions in the admiral universe, as admiral
package include \verb{#' @keywords deprecated} or \verb{#' @keywords superseded}
in the function's roxygen2 comments.
}
\examples{
get_admiral_deprecated()
get_admiral_superseded()
get_fns_with_keyword(package = "admiral", keyword = "superseded")
}
