% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiraldiscovery-package.R
\docType{package}
\name{admiraldiscovery-package}
\alias{admiraldiscovery}
\alias{admiraldiscovery-package}
\title{admiraldiscovery: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/admiraldiscovery/}
}

}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})

}
\keyword{internal}
