% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiraldiscovery-package.R
\docType{package}
\name{admiraldiscovery-package}
\alias{admiraldiscovery}
\alias{admiraldiscovery-package}
\title{admiraldiscovery: Discover admiral Functions}
\description{
Package exports a function that creates an interactive gt or reactable table to help users discover the admiral functions that are used to derive ADaM variables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/admiraldiscovery/}
  \item \url{https://github.com/pharmaverse/admiraldiscovery}
  \item Report bugs at \url{https://github.com/pharmaverse/admiraldiscovery/issues}
}

}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})

Authors:
\itemize{
  \item Amelia H. Tran \email{tran.amelia@gene.com} (\href{https://orcid.org/0000-0002-9472-0302}{ORCID})
}

}
\keyword{internal}
