% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{discovery}
\alias{discovery}
\title{Admiral Function Discovery}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 338 rows and 9 columns.
}
\usage{
discovery
}
\description{
This data set lists admiral functions and links them to ADaM data sets and
variables they help create.
}
\keyword{datasets}
