% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{interactive_discovery}
\alias{interactive_discovery}
\title{Interactive Admiral Function Discovery}
\usage{
interactive_discovery(type = c("gt", "reactable"))
}
\value{
interactive gt table
}
\description{
Interactively explore admiral functions and the ADaM data sets and variables
they help create.
}
\section{Demos}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{interactive_discovery()
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve--><div id="qvcmpguvya" class=".gt_table" style="padding-left:0px;padding-right:0px;padding-top:10px;padding-bottom:10px;overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
}}
\if{html}{\out{
<style>#qvcmpguvya table \{
  font-family: system-ui, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji';
  -webkit-font-smoothing: antialiased;
  -moz-osx-font-smoothing: grayscale;
\}

#qvcmpguvya thead, #qvcmpguvya tbody, #qvcmpguvya tfoot, #qvcmpguvya tr, #qvcmpguvya td, #qvcmpguvya th \{
  border-style: none;
\}

#qvcmpguvya p \{
  margin: 0;
  padding: 0;
\}

#qvcmpguvya .gt_table \{
  display: table;
  border-collapse: collapse;
  line-height: normal;
  margin-left: auto;
  margin-right: auto;
  color: #333333;
  font-size: 16px;
  font-weight: normal;
  font-style: normal;
  background-color: #FFFFFF;
  width: auto;
  border-top-style: solid;
  border-top-width: 2px;
  border-top-color: #A8A8A8;
  border-right-style: none;
  border-right-width: 2px;
  border-right-color: #D3D3D3;
  border-bottom-style: solid;
  border-bottom-width: 2px;
  border-bottom-color: #A8A8A8;
  border-left-style: none;
  border-left-width: 2px;
  border-left-color: #D3D3D3;
\}

#qvcmpguvya .gt_caption \{
  padding-top: 4px;
  padding-bottom: 4px;
\}

#qvcmpguvya .gt_title \{
  color: #333333;
  font-size: 125%;
  font-weight: initial;
  padding-top: 4px;
  padding-bottom: 4px;
  padding-left: 5px;
  padding-right: 5px;
  border-bottom-color: #FFFFFF;
  border-bottom-width: 0;
\}

#qvcmpguvya .gt_subtitle \{
  color: #333333;
  font-size: 85%;
  font-weight: initial;
  padding-top: 3px;
  padding-bottom: 5px;
  padding-left: 5px;
  padding-right: 5px;
  border-top-color: #FFFFFF;
  border-top-width: 0;
\}

#qvcmpguvya .gt_heading \{
  background-color: #FFFFFF;
  text-align: center;
  border-bottom-color: #FFFFFF;
  border-left-style: none;
  border-left-width: 1px;
  border-left-color: #D3D3D3;
  border-right-style: none;
  border-right-width: 1px;
  border-right-color: #D3D3D3;
\}

#qvcmpguvya .gt_bottom_border \{
  border-bottom-style: solid;
  border-bottom-width: 2px;
  border-bottom-color: #D3D3D3;
\}

#qvcmpguvya .gt_col_headings \{
  border-top-style: solid;
  border-top-width: 2px;
  border-top-color: #D3D3D3;
  border-bottom-style: solid;
  border-bottom-width: 2px;
  border-bottom-color: #D3D3D3;
  border-left-style: none;
  border-left-width: 1px;
  border-left-color: #D3D3D3;
  border-right-style: none;
  border-right-width: 1px;
  border-right-color: #D3D3D3;
\}

#qvcmpguvya .gt_col_heading \{
  color: #333333;
  background-color: #FFFFFF;
  font-size: 100%;
  font-weight: normal;
  text-transform: inherit;
  border-left-style: none;
  border-left-width: 1px;
  border-left-color: #D3D3D3;
  border-right-style: none;
  border-right-width: 1px;
  border-right-color: #D3D3D3;
  vertical-align: bottom;
  padding-top: 5px;
  padding-bottom: 6px;
  padding-left: 5px;
  padding-right: 5px;
  overflow-x: hidden;
\}

#qvcmpguvya .gt_column_spanner_outer \{
  color: #333333;
  background-color: #FFFFFF;
  font-size: 100%;
  font-weight: normal;
  text-transform: inherit;
  padding-top: 0;
  padding-bottom: 0;
  padding-left: 4px;
  padding-right: 4px;
\}

#qvcmpguvya .gt_column_spanner_outer:first-child \{
  padding-left: 0;
\}

#qvcmpguvya .gt_column_spanner_outer:last-child \{
  padding-right: 0;
\}

#qvcmpguvya .gt_column_spanner \{
  border-bottom-style: solid;
  border-bottom-width: 2px;
  border-bottom-color: #D3D3D3;
  vertical-align: bottom;
  padding-top: 5px;
  padding-bottom: 5px;
  overflow-x: hidden;
  display: inline-block;
  width: 100%;
\}

#qvcmpguvya .gt_spanner_row \{
  border-bottom-style: hidden;
\}

#qvcmpguvya .gt_group_heading \{
  padding-top: 8px;
  padding-bottom: 8px;
  padding-left: 5px;
  padding-right: 5px;
  color: #333333;
  background-color: #FFFFFF;
  font-size: 100%;
  font-weight: initial;
  text-transform: inherit;
  border-top-style: solid;
  border-top-width: 2px;
  border-top-color: #D3D3D3;
  border-bottom-style: solid;
  border-bottom-width: 2px;
  border-bottom-color: #D3D3D3;
  border-left-style: none;
  border-left-width: 1px;
  border-left-color: #D3D3D3;
  border-right-style: none;
  border-right-width: 1px;
  border-right-color: #D3D3D3;
  vertical-align: middle;
  text-align: left;
\}

#qvcmpguvya .gt_empty_group_heading \{
  padding: 0.5px;
  color: #333333;
  background-color: #FFFFFF;
  font-size: 100%;
  font-weight: initial;
  border-top-style: solid;
  border-top-width: 2px;
  border-top-color: #D3D3D3;
  border-bottom-style: solid;
  border-bottom-width: 2px;
  border-bottom-color: #D3D3D3;
  vertical-align: middle;
\}

#qvcmpguvya .gt_from_md > :first-child \{
  margin-top: 0;
\}

#qvcmpguvya .gt_from_md > :last-child \{
  margin-bottom: 0;
\}

#qvcmpguvya .gt_row \{
  padding-top: 8px;
  padding-bottom: 8px;
  padding-left: 5px;
  padding-right: 5px;
  margin: 10px;
  border-top-style: solid;
  border-top-width: 1px;
  border-top-color: #D3D3D3;
  border-left-style: none;
  border-left-width: 1px;
  border-left-color: #D3D3D3;
  border-right-style: none;
  border-right-width: 1px;
  border-right-color: #D3D3D3;
  vertical-align: middle;
  overflow-x: hidden;
\}

#qvcmpguvya .gt_stub \{
  color: #333333;
  background-color: #FFFFFF;
  font-size: 100%;
  font-weight: initial;
  text-transform: inherit;
  border-right-style: solid;
  border-right-width: 2px;
  border-right-color: #D3D3D3;
  padding-left: 5px;
  padding-right: 5px;
\}

#qvcmpguvya .gt_stub_row_group \{
  color: #333333;
  background-color: #FFFFFF;
  font-size: 100%;
  font-weight: initial;
  text-transform: inherit;
  border-right-style: solid;
  border-right-width: 2px;
  border-right-color: #D3D3D3;
  padding-left: 5px;
  padding-right: 5px;
  vertical-align: top;
\}

#qvcmpguvya .gt_row_group_first td \{
  border-top-width: 2px;
\}

#qvcmpguvya .gt_row_group_first th \{
  border-top-width: 2px;
\}

#qvcmpguvya .gt_summary_row \{
  color: #333333;
  background-color: #FFFFFF;
  text-transform: inherit;
  padding-top: 8px;
  padding-bottom: 8px;
  padding-left: 5px;
  padding-right: 5px;
\}

#qvcmpguvya .gt_first_summary_row \{
  border-top-style: solid;
  border-top-color: #D3D3D3;
\}

#qvcmpguvya .gt_first_summary_row.thick \{
  border-top-width: 2px;
\}

#qvcmpguvya .gt_last_summary_row \{
  padding-top: 8px;
  padding-bottom: 8px;
  padding-left: 5px;
  padding-right: 5px;
  border-bottom-style: solid;
  border-bottom-width: 2px;
  border-bottom-color: #D3D3D3;
\}

#qvcmpguvya .gt_grand_summary_row \{
  color: #333333;
  background-color: #FFFFFF;
  text-transform: inherit;
  padding-top: 8px;
  padding-bottom: 8px;
  padding-left: 5px;
  padding-right: 5px;
\}

#qvcmpguvya .gt_first_grand_summary_row \{
  padding-top: 8px;
  padding-bottom: 8px;
  padding-left: 5px;
  padding-right: 5px;
  border-top-style: double;
  border-top-width: 6px;
  border-top-color: #D3D3D3;
\}

#qvcmpguvya .gt_last_grand_summary_row_top \{
  padding-top: 8px;
  padding-bottom: 8px;
  padding-left: 5px;
  padding-right: 5px;
  border-bottom-style: double;
  border-bottom-width: 6px;
  border-bottom-color: #D3D3D3;
\}

#qvcmpguvya .gt_striped \{
  background-color: rgba(128, 128, 128, 0.05);
\}

#qvcmpguvya .gt_table_body \{
  border-top-style: solid;
  border-top-width: 2px;
  border-top-color: #D3D3D3;
  border-bottom-style: solid;
  border-bottom-width: 2px;
  border-bottom-color: #D3D3D3;
\}

#qvcmpguvya .gt_footnotes \{
  color: #333333;
  background-color: #FFFFFF;
  border-bottom-style: none;
  border-bottom-width: 2px;
  border-bottom-color: #D3D3D3;
  border-left-style: none;
  border-left-width: 2px;
  border-left-color: #D3D3D3;
  border-right-style: none;
  border-right-width: 2px;
  border-right-color: #D3D3D3;
\}

#qvcmpguvya .gt_footnote \{
  margin: 0px;
  font-size: 90%;
  padding-top: 4px;
  padding-bottom: 4px;
  padding-left: 5px;
  padding-right: 5px;
\}

#qvcmpguvya .gt_sourcenotes \{
  color: #333333;
  background-color: #FFFFFF;
  border-bottom-style: none;
  border-bottom-width: 2px;
  border-bottom-color: #D3D3D3;
  border-left-style: none;
  border-left-width: 2px;
  border-left-color: #D3D3D3;
  border-right-style: none;
  border-right-width: 2px;
  border-right-color: #D3D3D3;
\}

#qvcmpguvya .gt_sourcenote \{
  font-size: 90%;
  padding-top: 4px;
  padding-bottom: 4px;
  padding-left: 5px;
  padding-right: 5px;
\}

#qvcmpguvya .gt_left \{
  text-align: left;
\}

#qvcmpguvya .gt_center \{
  text-align: center;
\}

#qvcmpguvya .gt_right \{
  text-align: right;
  font-variant-numeric: tabular-nums;
\}

#qvcmpguvya .gt_font_normal \{
  font-weight: normal;
\}

#qvcmpguvya .gt_font_bold \{
  font-weight: bold;
\}

#qvcmpguvya .gt_font_italic \{
  font-style: italic;
\}

#qvcmpguvya .gt_super \{
  font-size: 65%;
\}

#qvcmpguvya .gt_footnote_marks \{
  font-size: 75%;
  vertical-align: 0.4em;
  position: initial;
\}

#qvcmpguvya .gt_asterisk \{
  font-size: 100%;
  vertical-align: 0;
\}

#qvcmpguvya .gt_indent_1 \{
  text-indent: 5px;
\}

#qvcmpguvya .gt_indent_2 \{
  text-indent: 10px;
\}

#qvcmpguvya .gt_indent_3 \{
  text-indent: 15px;
\}

#qvcmpguvya .gt_indent_4 \{
  text-indent: 20px;
\}

#qvcmpguvya .gt_indent_5 \{
  text-indent: 25px;
\}
</style>
}}
\if{html}{\out{
<div id="qvcmpguvya" class="reactable html-widget " style="width:auto;height:auto;"></div>
<script type="application/json" data-for="qvcmpguvya">\{"x":\{"tag":\{"name":"Reactable","attribs":\{"data":\{"dataset":["ADSL","ADVS"],"dataset_type":["ADSL","BDS"],"variable":["AGE","PARAMCD:BMI"],"variable_label":["Age","Body Mass Index (kg/m^2)"],"function_link":["[`admiral::compute_age_years()`](https://pharmaverse.github.io/admiral/v0.11.1/reference/compute_age_years.html)","[`admiral::derive_param_bmi()`](https://pharmaverse.github.io/admiral/v0.11.1/reference/derive_param_bmi.html)"],"adam_ig_reference":["ADaMIG v1.3. Table 3.2.2 ADSL Subject Demographics Variables",null]\},"columns":[\{"id":"dataset","name":"Data Set","type":"character","style":"function(rowInfo, colInfo) \{\\nconst rowIndex = rowInfo.index + 1\\n\}","cell":["ADSL","ADVS"],"html":true,"align":"left","headerStyle":\{"font-weight":"normal"\}\},\{"id":"dataset_type","name":"Data Set Type","type":"character","style":"function(rowInfo, colInfo) \{\\nconst rowIndex = rowInfo.index + 1\\n\}","cell":["ADSL","BDS"],"html":true,"align":"left","headerStyle":\{"font-weight":"normal"\}\},\{"id":"variable","name":"Variable","type":"character","style":"function(rowInfo, colInfo) \{\\nconst rowIndex = rowInfo.index + 1\\n\}","cell":["AGE","PARAMCD:BMI"],"html":true,"align":"left","headerStyle":\{"font-weight":"normal"\}\},\{"id":"variable_label","name":"Variable Label","type":"character","style":"function(rowInfo, colInfo) \{\\nconst rowIndex = rowInfo.index + 1\\n\}","cell":["Age","Body Mass Index (kg/m^2)"],"html":true,"align":"left","headerStyle":\{"font-weight":"normal"\}\},\{"id":"function_link","name":"Function","type":["glue","character"],"style":"function(rowInfo, colInfo) \{\\nconst rowIndex = rowInfo.index + 1\\n\}","cell":["<div class='gt_from_md'><p><a href=\\"https://pharmaverse.github.io/admiral/v0.11.1/reference/compute_age_years.html\\"><code>admiral::compute_age_years()<\\/code><\\/a><\\/p>\\n<\\/div>","<div class='gt_from_md'><p><a href=\\"https://pharmaverse.github.io/admiral/v0.11.1/reference/derive_param_bmi.html\\"><code>admiral::derive_param_bmi()<\\/code><\\/a><\\/p>\\n<\\/div>"],"html":true,"align":"left","headerStyle":\{"font-weight":"normal"\}\},\{"id":"adam_ig_reference","name":"ADaM Implmentation Guide Reference","type":"character","style":"function(rowInfo, colInfo) \{\\nconst rowIndex = rowInfo.index + 1\\n\}","cell":["ADaMIG v1.3. Table 3.2.2 ADSL Subject Demographics Variables","<br />"],"html":true,"align":"left","headerStyle":\{"font-weight":"normal"\}\}],"resizable":true,"filterable":true,"searchable":true,"defaultPageSize":25,"showPageSizeOptions":true,"pageSizeOptions":[10,25,50,100],"paginationType":"numbers","showPagination":true,"showPageInfo":true,"minRows":1,"highlight":true,"compact":true,"showSortable":true,"height":"auto","theme":\{"color":"#333333","backgroundColor":"#FFFFFF","stripedColor":"rgba(128,128,128,0.05)","style":\{"fontFamily":"system-ui, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif"\},"headerStyle":\{"borderTopStyle":"solid","borderTopWidth":"2px","borderTopColor":"#D3D3D3","borderBottomStyle":"solid","borderBottomWidth":"2px","borderBottomColor":"#D3D3D3"\}\},"elementId":"qvcmpguvya","dataKey":"e7d1c9491e7131e04d8d4189e9d8f1b1"\},"children":[]\},"class":"reactR_markup"\},"evals":["tag.attribs.columns.0.style","tag.attribs.columns.1.style","tag.attribs.columns.2.style","tag.attribs.columns.3.style","tag.attribs.columns.4.style","tag.attribs.columns.5.style"],"jsHooks":[]\}</script>
</div><!--/html_preserve-->
}}
}

