% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{interactive_discovery}
\alias{interactive_discovery}
\title{Interactive Admiral Function Discovery}
\usage{
interactive_discovery(type = c("gt", "reactable"))
}
\arguments{
\item{type}{type of table to return}
}
\value{
interactive gt table
}
\description{
Interactively explore admiral functions and the ADaM data sets and variables
they help create.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
interactive_discovery("gt")
interactive_discovery("reactable")
\dontshow{\}) # examplesIf}
}
