% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{presentation_archive}
\alias{presentation_archive}
\title{Presentation Archive}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 20 rows and 5 columns.
}
\usage{
presentation_archive
}
\description{
This data set contains information and links to resources of recent presentations
about admiral-related topics.
}
\keyword{datasets}
